+++ web/static/react/assets/ColdStartInterviewPage-DTY4Fnj5.js
@@ -1,0 +1,1 @@
+import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as t}from"./LegalSkillsProfilePage-dFSWuwOq.js";import"./button-HFB6Zh4Y.js";import"./index-DM1xmz2w.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-CVFDk9Jl.js";import"./IusPageShell-BGy80zt2.js";import"./api-B69c6ny4.js";function n(){return r.jsx(t,{})}export{n as ColdStartInterviewPage};